backend/requirements-backend.txt
fastapi
uvicorn
python-docx
PyMuPDF
python-dotenv
openai
scikit-learn
streamlit
requests
fpdf